+++ CHANGELOG.md
@@ -6,6 +6,7 @@ _released 00 September 2025_
 
 - Features:
   - View the full date for a day by hovering over the number in the calendar.
+  - When an assignment is graded, the grade is shown in the assignment popup.
 - Bug Fixes:
   - Fixes some styling bugs in the toolbar.
   - Assignment boxes keep their focus styling (lighter background color) when